refactor: 코드리뷰반영, view클래스 분리에 다른 패키지 정리
