+++ src/main/java/lotto/view/BuildResponseMessage.java
@@ -1,0 +1,6 @@
+package lotto.view;
+
+@FunctionalInterface
+public interface BuildResponseMessage {
+    String toMessage();
+}
