--- src/main/java/lotto/view/ResponseView.java
@@ -1,10 +1,0 @@
-package lotto.view;
-
-//@FunctionalInterface
-public interface ResponseView {
-
-    @Deprecated
-    String toView();
-
-    void show();
-}
