+++ src/main/java/lotto/view/input/PaymentAmountInputView.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.view.input;
 
 import java.util.Scanner;
