+++ src/main/java/lotto/view/response/BuyResultResponseView.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.view.response;
 
 public class BuyResultResponseView {
