+++ src/main/java/lotto/view/response/GeneratedLottoResponseView.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.view.response;
 
 
 import java.util.List;
