+++ src/main/java/lotto/view/response/LottoYieldResponseView.java
@@ -1,8 +1,9 @@
-package lotto.view;
+package lotto.view.response;
 
 import lotto.LottoYield;
+import lotto.view.BuildResponseMessage;
 
-public class LottoYieldResponseView implements ResponseView {
+public class LottoYieldResponseView implements BuildResponseMessage {
     private static final String FORMAT = "총 수익률은 %.2f입니다.";
     private static final String WARNING = "(기준이 1이기 때문에 결과적으로 손해라는 의미임)";
 
@@ -13,7 +14,7 @@ public class LottoYieldResponseView implements ResponseView {
     }
 
     @Override
-    public String toView() {
+    public String toMessage() {
         double yield = lottoYield.calculate();
 
         String view = String.format(FORMAT, yield);
